add field name for user after first time login
