--- app/src/main/java/com/tugasakhir/resq/korban/MasukActivity.kt
@@ -1,36 +1,0 @@
-package com.tugasakhir.resq.korban
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.ActionBar
-import androidx.appcompat.app.AppCompatActivity
-import com.tugasakhir.resq.R
-import kotlinx.android.synthetic.main.activity_korban_masuk.*
-
-class MasukActivity : AppCompatActivity() {
-
-    private lateinit var actionBar : ActionBar
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_korban_masuk)
-
-        actionBar = this.supportActionBar!!
-        actionBar.setHomeAsUpIndicator(R.mipmap.ic_logo_round)
-        actionBar.setDisplayHomeAsUpEnabled(true)
-        actionBar.title = "MASUK"
-        actionBar.elevation = 0F
-
-        button_signin_tosignup.setOnClickListener {
-            val intent = Intent(this, BuatAkunActivity::class.java)
-            startActivity(intent)
-        }
-
-        button_signin_continue.setOnClickListener {
-            val intent = Intent(this, OTPActivity::class.java)
-            startActivity(intent)
-        }
-
-    }
-
-}
